+++ src/core/Solutions/index.ts
@@ -1,3 +1,10 @@
-export * from "./model";
+export {
+    Solution,
+    SolutionType,
+    SolutionItem,
+    SolutionFolder,
+    SolutionProject,
+    SolutionProjectType,
+} from "./model";
 export * from "./SolutionFactory";
 export * from "../../SolutionFinder";